requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
Werkzeug==2.0.3
google-generativeai==0.3.1
python-dotenv==1.0.0

